Rename the header component to more properly match its actually functionality
