+++ client/src/containers/Footer/index.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 // View
-import HeaderView from '../../components/Header';
+import FooterView from '../../components/Footer';
 
 // actions
 import { hideModal, dispatchShowModal } from '../../reducers/Modal/actions';
@@ -11,18 +11,16 @@ import { logOutUser } from '../../reducers/User/actions';
 import { showAllBuoys } from '../../reducers/Data/actions';
 
 /**
- * @description Header handles logic for buttons on the NavBar and their text
+ * @description Footer handles logic for buttons on the NavBar and their text
  *
  * @param {object} props - Component props
- * @param {boolean} props.dataError - Error message for data actions
- * @param {boolean} props.userError - Error message for user actions
  * @param {function} props.dispatchHideModal - Method for dispatching an action to hide the modal.
  * @param {function} props.dispatchShowModal - Method for dispatching an action to show a specific modal.
  * @param {function} props.dispatchLogoutUser - Method for dispatching an action to kill a user session.
  *
- * @returns {<Header />}
+ * @returns {<Footer />}
  */
-class Header extends PureComponent {
+class Footer extends PureComponent {
   constructor(props) {
     super(props);
 
@@ -60,11 +58,11 @@ class Header extends PureComponent {
       loggedOutStyle,
     };
 
-    return <HeaderView {...props} />;
+    return <FooterView {...props} />;
   }
 }
 
-Header.propTypes = {
+Footer.propTypes = {
   // Props
   loggedIn: PropTypes.bool,
 
@@ -72,7 +70,7 @@ Header.propTypes = {
   dispatchShowModal: PropTypes.func.isRequired,
 };
 
-Header.defaultProps = {
+Footer.defaultProps = {
   loggedIn: false,
 };
 
@@ -93,4 +91,4 @@ const mapDispatchToProps = dispatch => ({
   },
 });
 
-export default connect(mapStateToProps, mapDispatchToProps)(Header);
+export default connect(mapStateToProps, mapDispatchToProps)(Footer);
