+++ client/src/containers/index.js
@@ -1,6 +1,6 @@
 export { default as App } from './App';
 export { default as ErrorBoundary } from './ErrorBoundary';
-export { default as Header } from './Header';
+export { default as Footer } from './Footer';
 export { default as Modal } from './Modal';
 export { default as BuoyList } from './BuoyList';
 export { default as Buoy } from './Buoy';
