+++ client/src/styles/Dashboard/Dashboard.scss
@@ -19,3 +19,18 @@
   width: 100%;
   padding-top: 10px;
 }
+
+.sp-header-error {
+  background-color: rgba(210, 77, 87, 1);
+  padding: 10px;
+  color: white;
+  width: 100%;
+  text-align: center;
+  margin-bottom: 10px;
+}
+
+.sp-error-msg {
+  font-family: 'Montserrat', sans-serif;
+  font-size: 13px;
+  font-weight: 300;
+}
